ref earlier release note point of confusion
Release notes a few months ago talked about Report Totals showing all data, whether Unspecified displayed or not. Now we use terms Table Total and Grand Total. We should provide a link or comment to clarify that the older notes apply to what is now called Grand Total. As an experienced user looking for info on Unspecified, I saw the older notes about Report Totals and believed it would apply to "Table Totals" which is inaccurate.